FIX : Carousel, ProgressBox 컴포넌트 분리
