+++ components/newHome/Services/ProgressBox.module.scss
@@ -1,0 +1,121 @@
+@use "../../../styles/position";
+@import "../../../styles/lib";
+
+$carousel-length: 3;
+
+.progress-box {
+  margin-top: px2vh(51px);
+  display: flex;
+  align-items: center;
+  gap: px2vw(6px);
+
+  > .bar-wrapper {
+    position: relative;
+    > .bar {
+      position: absolute;
+      height: 3px;
+      border-radius: 1.5px;
+    }
+    > .background-bar {
+      width: 202px;
+      background-color: #ede2e0;
+    }
+    > .foreground-bar {
+      background-color: #f0745f;
+    }
+  }
+
+  > .index-wrapper {
+    margin-left: calc(6px + 202px);
+    display: flex;
+    align-items: center;
+    gap: 3px;
+
+    > .index {
+      width: 32.58px;
+      padding: 10px;
+      font-size: 18px;
+      font-weight: 600;
+      line-height: 160%;
+      letter-spacing: 0.04em;
+      color: #e0e0e0;
+    }
+    > .current-index {
+      color: #f0745f;
+    }
+
+    > .divider {
+      width: 1px;
+      height: 15px;
+      background-color: #e0e0e0;
+      border-radius: 0.5px;
+    }
+  }
+
+  > .play-button {
+    display: flex;
+    align-items: center;
+    position: relative;
+    height: 29px;
+    background-color: transparent;
+    border: none;
+    padding: 0;
+    margin: 0;
+    background-repeat: no-repeat;
+    background-position: center;
+    &:hover {
+      cursor: pointer;
+    }
+    > .border {
+      position: absolute;
+      width: 29px;
+      height: 29px;
+      border: 1px solid #e0e0e0;
+      border-radius: 14.5px;
+    }
+    > .foreground-border {
+      @include position.flex-center;
+      z-index: 1;
+      > svg {
+        stroke-dasharray: 88;
+        stroke-dashoffset: 88;
+        transform: rotate(-90deg);
+        animation-name: ani;
+        animation-duration: 3s;
+        animation-timing-function: linear;
+        animation-delay: 0s;
+        animation-play-state: paused;
+        animation-iteration-count: infinite;
+      }
+      @keyframes ani {
+        0% {
+          stroke-dashoffset: 88;
+        }
+        100% {
+          stroke-dashoffset: 0;
+        }
+      }
+    }
+  }
+  > .play-button.playing {
+    width: 29px;
+    background-image: url("/static/images/newHome/stop_button_icon.svg");
+    > .foreground-border {
+      > svg {
+        animation-play-state: running;
+      }
+    }
+  }
+  > .play-button.stopped {
+    width: 31px;
+    background-image: url("/static/images/newHome/play_button_icon.svg");
+    > svg {
+      animation-play-state: paused;
+    }
+  }
+}
+@for $i from 1 through $carousel-length {
+  > .progress-box.state-#{$i} > .bar-wrapper > .foreground-bar {
+    width: calc((202px * $i) / $carousel-length);
+  }
+}
