+++ components/newHome/Services/Services.module.scss
@@ -2,8 +2,6 @@
 @use "../../../styles/position";
 @import "../../../styles/lib";
 
-$carousel-length: 3;
-
 .container {
   width: 100vw;
   height: 100vh;
@@ -53,190 +51,6 @@ $carousel-length: 3;
             color: #a4a4a4;
             height: px2vh(25px);
             line-height: px2vh(25px);
-          }
-        }
-
-        > .progress-box {
-          margin-top: px2vh(51px);
-          display: flex;
-          align-items: center;
-          gap: px2vw(6px);
-
-          > .bar-wrapper {
-            position: relative;
-            > .bar {
-              position: absolute;
-              height: 3px;
-              border-radius: 1.5px;
-            }
-            > .background-bar {
-              width: 202px;
-              background-color: #ede2e0;
-            }
-            > .foreground-bar {
-              background-color: #f0745f;
-            }
-          }
-
-          > .index-wrapper {
-            margin-left: calc(6px + 202px);
-            display: flex;
-            align-items: center;
-            gap: 3px;
-
-            > .index {
-              width: 32.58px;
-              padding: 10px;
-              font-size: 18px;
-              font-weight: 600;
-              line-height: 160%;
-              letter-spacing: 0.04em;
-              color: #e0e0e0;
-            }
-            > .current-index {
-              color: #f0745f;
-            }
-
-            > .divider {
-              width: 1px;
-              height: 15px;
-              background-color: #e0e0e0;
-              border-radius: 0.5px;
-            }
-          }
-
-          > .play-button {
-            display: flex;
-            align-items: center;
-            position: relative;
-            height: 29px;
-            background-color: transparent;
-            border: none;
-            padding: 0;
-            margin: 0;
-            background-repeat: no-repeat;
-            background-position: center;
-            &:hover {
-              cursor: pointer;
-            }
-            > .border {
-              position: absolute;
-              width: 29px;
-              height: 29px;
-              border: 1px solid #e0e0e0;
-              border-radius: 14.5px;
-            }
-            > .foreground-border {
-              @include position.flex-center;
-              z-index: 1;
-              > svg {
-                stroke-dasharray: 88;
-                stroke-dashoffset: 88;
-                transform: rotate(-90deg);
-                animation-name: ani;
-                animation-duration: 3s;
-                animation-timing-function: linear;
-                animation-delay: 0s;
-                animation-play-state: paused;
-                animation-iteration-count: infinite;
-              }
-              @keyframes ani {
-                0% {
-                  stroke-dashoffset: 88;
-                }
-                100% {
-                  stroke-dashoffset: 0;
-                }
-              }
-            }
-          }
-          > .play-button.playing {
-            width: 29px;
-            background-image: url("/static/images/newHome/stop_button_icon.svg");
-            > .foreground-border {
-              > svg {
-                animation-play-state: running;
-              }
-            }
-          }
-          > .play-button.stopped {
-            width: 31px;
-            background-image: url("/static/images/newHome/play_button_icon.svg");
-            > svg {
-              animation-play-state: paused;
-            }
-          }
-        }
-        @for $i from 1 through $carousel-length {
-          > .progress-box.state-#{$i} > .bar-wrapper > .foreground-bar {
-            width: calc((202px * $i) / $carousel-length);
-          }
-        }
-      }
-      > .carousel {
-        position: relative;
-        margin-top: calc(75px + px2vh(130px));
-        display: flex;
-        align-items: center;
-        gap: px2vw(75px);
-
-        > .carousel-button {
-          padding: 0;
-          border: none;
-          background-color: transparent;
-          cursor: pointer;
-        }
-        > .carousel-frame {
-          width: 528px;
-          height: 584px;
-          background-color: transparent;
-          overflow-x: hidden;
-          // overflow-y: hidden;
-          > .carousel-items-container {
-            display: flex;
-            transition: 0.5s;
-            gap: 100px;
-            > .carousel-item {
-              flex: none;
-              display: flex;
-              width: 528px;
-              height: 582px;
-              align-items: center;
-              gap: px2vw(54px);
-              > .left {
-                > .service-icon {
-                  display: block;
-                }
-                > .service-name {
-                  font-size: 28px;
-                  font-weight: bold;
-                  letter-spacing: 0.04em;
-                  line-height: 160%;
-                  color: #323232;
-                  text-align: center;
-                  margin: -2px 0;
-                }
-                > .description {
-                  font-size: 16px;
-                  letter-spacing: 0.04em;
-                  line-height: 160%;
-                  color: #737373;
-                  text-align: center;
-                }
-              }
-              > .phone {
-                display: block;
-              }
-            }
-          }
-
-          @for $i from 1 through $carousel-length {
-            > .carousel-items-container.state-#{$i} {
-              transform: translateX(calc(628px) * -($i - 1));
-            }
-          }
-          > .carousel-items-container.state-1 {
-            transition: 1s;
           }
         }
       }
